morse code converted text, audio, GUI widgets position relative
The text conversio to morse code is working perfectly fine and the output is clearly visible on the GUI, the widgets of GUI are also relatively positioned so that the orientation does not get spoiled on changing the size of the application window. The audio is also working perfectly fine on clicking the audio button.